Start refactoring Scene tests into separate modules
