+++ satpy/tests/test_scene.py
@@ -39,11 +39,8 @@ from satpy.tests.utils import (
     default_id_keys_config,
     make_cid,
     make_dataid,
-    make_dsq,
     spy_decorator,
 )
-
-TEST_ETC_DIR = os.path.join(os.path.dirname(__file__), 'etc')
 
 
 def _check_comp19_deps_are_loaded(scene):
@@ -57,17 +54,9 @@ def _check_comp19_deps_are_loaded(scene):
         assert any(x['name'] == name for x in loaded_ids)
 
 
+@pytest.mark.usefixtures("include_test_etc")
 class TestScene:
     """Test the scene class."""
-
-    def setup_method(self):
-        """Set config_path to point to test 'etc' directory."""
-        self.old_config_path = satpy.config.get('config_path')
-        satpy.config.set(config_path=[TEST_ETC_DIR])
-
-    def teardown_method(self):
-        """Restore previous 'config_path' setting."""
-        satpy.config.set(config_path=self.old_config_path)
 
     def test_init(self):
         """Test scene initialization."""
@@ -242,11 +231,10 @@ class TestScene:
             assert reader_kwargs == rinit.mock.call_args[1]['fh_kwargs']
             del scene
 
-    def test_create_multiple_reader_different_kwargs(self):
+    def test_create_multiple_reader_different_kwargs(self, include_test_etc):
         """Test passing different kwargs to different readers."""
         from satpy.readers import load_reader
-        with satpy.config.set(config_path=[TEST_ETC_DIR]), \
-             mock.patch.object(satpy.readers, 'load_reader', wraps=load_reader) as lr:
+        with mock.patch.object(satpy.readers, 'load_reader', wraps=load_reader) as lr:
             Scene(filenames={"fake1_1ds": ["fake1_1ds_1.txt"],
                              "fake2_1ds": ["fake2_1ds_1.txt"]},
                   reader_kwargs={
@@ -254,8 +242,8 @@ class TestScene:
                       "fake2_1ds": {"mouth": "varallo"}
                   })
             lr.assert_has_calls([
-                    mock.call([os.path.join(TEST_ETC_DIR, 'readers', 'fake1_1ds.yaml')], mouth="omegna"),
-                    mock.call([os.path.join(TEST_ETC_DIR, 'readers', 'fake2_1ds.yaml')], mouth="varallo")])
+                    mock.call([os.path.join(include_test_etc, 'readers', 'fake1_1ds.yaml')], mouth="omegna"),
+                    mock.call([os.path.join(include_test_etc, 'readers', 'fake2_1ds.yaml')], mouth="varallo")])
 
     def test_iter(self):
         """Test iteration over the scene."""
@@ -840,80 +828,9 @@ class TestFinestCoarsestArea:
         assert coarse_area2 is course_area1
 
 
-class TestSceneAvailableDatasets:
-    """Test the Scene's handling of various dependencies."""
-
-    def setup_method(self):
-        """Set config_path to point to test 'etc' directory."""
-        self.old_config_path = satpy.config.get('config_path')
-        satpy.config.set(config_path=[TEST_ETC_DIR])
-
-    def teardown_method(self):
-        """Restore previous 'config_path' setting."""
-        satpy.config.set(config_path=self.old_config_path)
-
-    def test_all_datasets_one_reader(self):
-        """Test all datasets for one reader."""
-        scene = Scene(filenames=['fake1_1.txt'],
-                      reader='fake1')
-        id_list = scene.all_dataset_ids()
-        # 20 data products + 6 lon/lat products
-        num_reader_ds = 21 + 6
-        assert len(id_list) == num_reader_ds
-        id_list = scene.all_dataset_ids(composites=True)
-        assert len(id_list) == num_reader_ds + 33
-
-    def test_all_datasets_multiple_reader(self):
-        """Test all datasets for multiple readers."""
-        scene = Scene(filenames={'fake1_1ds': ['fake1_1ds_1.txt'],
-                                 'fake2_1ds': ['fake2_1ds_1.txt']})
-        id_list = scene.all_dataset_ids()
-        assert len(id_list) == 2
-        id_list = scene.all_dataset_ids(composites=True)
-        # ds1 and ds2 => 2
-        # composites that use these two datasets => 11
-        assert len(id_list) == 2 + 11
-
-    def test_available_datasets_one_reader(self):
-        """Test the available datasets for one reader."""
-        scene = Scene(filenames=['fake1_1ds_1.txt'],
-                      reader='fake1_1ds')
-        id_list = scene.available_dataset_ids()
-        assert len(id_list) == 1
-        id_list = scene.available_dataset_ids(composites=True)
-        # ds1, comp1, comp14, comp16, static_image, comp26
-        assert len(id_list) == 6
-
-    def test_available_composite_ids_missing_available(self):
-        """Test available_composite_ids when a composites dep is missing."""
-        scene = Scene(filenames=['fake1_1ds_1.txt'],
-                      reader='fake1_1ds')
-        assert 'comp2' not in scene.available_composite_names()
-
-    def test_available_composites_known_versus_all(self):
-        """Test available_composite_ids when some datasets aren't available."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1',
-                      reader_kwargs={"not_available": ["ds2", "ds3"]})
-        all_comps = scene.all_composite_names()
-        avail_comps = scene.available_composite_names()
-        # there should always be more known composites than available composites
-        assert len(all_comps) > len(avail_comps)
-        for not_avail_comp in ("comp2", "comp3"):
-            assert not_avail_comp in all_comps
-            assert not_avail_comp not in avail_comps
-
-
+@pytest.mark.usefixtures("include_test_etc")
 class TestSceneSerialization:
     """Test the Scene serialization."""
-
-    def setup_method(self):
-        """Set config_path to point to test 'etc' directory."""
-        self.old_config_path = satpy.config.get('config_path')
-        satpy.config.set(config_path=[TEST_ETC_DIR])
-
-    def teardown_method(self):
-        """Restore previous 'config_path' setting."""
-        satpy.config.set(config_path=self.old_config_path)
 
     def test_serialization_with_readers_and_data_arr(self):
         """Test that dask can serialize a Scene with readers."""
@@ -926,703 +843,9 @@ class TestSceneSerialization:
         assert scene.all_dataset_ids == scene.all_dataset_ids
 
 
-class TestSceneLoading:
-    """Test the Scene objects `.load` method."""
-
-    def setup_method(self):
-        """Set config_path to point to test 'etc' directory."""
-        self.old_config_path = satpy.config.get('config_path')
-        satpy.config.set(config_path=[TEST_ETC_DIR])
-
-    def teardown_method(self):
-        """Restore previous 'config_path' setting."""
-        satpy.config.set(config_path=self.old_config_path)
-
-    def test_load_str(self):
-        """Test passing a string to Scene.load."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        pytest.raises(TypeError, scene.load, 'ds1')
-
-    def test_load_no_exist(self):
-        """Test loading a dataset that doesn't exist."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        pytest.raises(KeyError, scene.load, ['im_a_dataset_that_doesnt_exist'])
-
-    def test_load_no_exist2(self):
-        """Test loading a dataset that doesn't exist then another load."""
-        from satpy.readers.yaml_reader import FileYAMLReader
-        load_mock = spy_decorator(FileYAMLReader.load)
-        with mock.patch.object(FileYAMLReader, 'load', load_mock):
-            lmock = load_mock.mock
-            scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-            scene.load(['ds9_fail_load'])
-            loaded_ids = list(scene._datasets.keys())
-            assert len(loaded_ids) == 0
-            lmock.assert_called_once_with(
-                set([make_dataid(name='ds9_fail_load', wavelength=(1.0, 1.1, 1.2))]))
-
-            scene.load(['ds1'])
-            loaded_ids = list(scene._datasets.keys())
-            assert lmock.call_count == 2
-            # most recent call should have only been ds1
-            lmock.assert_called_with(set([
-                make_dataid(name='ds1', resolution=250,
-                            calibration='reflectance', modifiers=tuple())]))
-            assert len(loaded_ids) == 1
-
-    def test_load_ds1_no_comps(self):
-        """Test loading one dataset with no loaded compositors."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_dataid(name='ds1', resolution=250, calibration='reflectance', modifiers=tuple())
-
-    def test_load_ds1_load_twice(self):
-        """Test loading one dataset with no loaded compositors."""
-        from satpy.readers.yaml_reader import FileYAMLReader
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_dataid(name='ds1', resolution=250, calibration='reflectance', modifiers=tuple())
-
-        load_mock = spy_decorator(FileYAMLReader.load)
-        with mock.patch.object(FileYAMLReader, 'load', load_mock):
-            lmock = load_mock.mock
-            scene.load(['ds1'])
-            loaded_ids = list(scene._datasets.keys())
-            assert len(loaded_ids) == 1
-            assert loaded_ids[0] == make_dataid(name='ds1',
-                                                resolution=250,
-                                                calibration='reflectance',
-                                                modifiers=tuple())
-            assert not lmock.called, ("Reader.load was called again when "
-                                      "loading something that's already "
-                                      "loaded")
-
-    def test_load_ds1_unknown_modifier(self):
-        """Test loading one dataset with no loaded compositors."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        pytest.raises(KeyError, scene.load,
-                      [make_dataid(name='ds1', modifiers=('_fake_bad_mod_',))])
-
-    def test_load_ds4_cal(self):
-        """Test loading a dataset that has two calibration variations."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds4'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['calibration'] == 'reflectance'
-
-    def test_load_ds5_best_resolution(self):
-        """Test loading a dataset has multiple resolutions available."""
-        scene = Scene(filenames=['fake1_1.txt', 'fake1_highres_1.txt'], reader='fake1')
-        scene.load(['ds5'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'ds5'
-        assert loaded_ids[0]['resolution'] == 250
-
-    def test_load_ds5_multiple_resolution(self):
-        """Test loading a dataset has multiple resolutions available with different resolutions."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds5'], resolution=1000)
-        scene.load(['ds5'], resolution=500)
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        assert loaded_ids[0]['name'] == 'ds5'
-        assert loaded_ids[0]['resolution'] == 500
-        assert loaded_ids[1]['name'] == 'ds5'
-        assert loaded_ids[1]['resolution'] == 1000
-
-    def test_load_ds5_resolution_list(self):
-        """Test loading a dataset has multiple resolutions available with different resolutions."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds5'], resolution=[500, 1000])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'ds5'
-        assert loaded_ids[0]['resolution'] == 500
-
-    def test_load_ds5_empty_modifiers(self):
-        """Test loading a dataset has multiple resolutions available with different resolutions."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load([make_dsq(name='ds5', modifiers=tuple())])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'ds5'
-
-    def test_load_ds5_missing_best_resolution(self):
-        """Test loading a dataset that has multiple resolutions but the best isn't available."""
-        # only the 500m is available
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds5'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'ds5'
-        assert loaded_ids[0]['resolution'] == 500
-
-    def test_load_ds6_wl(self):
-        """Test loading a dataset by wavelength."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load([0.22])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'ds6'
-
-    def test_load_ds9_fail_load(self):
-        """Test loading a dataset that will fail during load."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds9_fail_load'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 0
-
-    def test_load_comp1(self):
-        """Test loading a composite with one required prereq."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp1'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp1')
-
-    def test_load_comp4(self):
-        """Test loading a composite that depends on a composite."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp4'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp4')
-
-    def test_load_multiple_resolutions(self):
-        """Test loading a dataset has multiple resolutions available with different resolutions."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        comp25 = make_cid(name='comp25', resolution=1000)
-        scene[comp25] = xr.DataArray([], attrs={'name': 'comp25', 'resolution': 1000})
-        scene.load(['comp25'], resolution=500)
-
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        assert loaded_ids[0]['name'] == 'comp25'
-        assert loaded_ids[0]['resolution'] == 500
-        assert loaded_ids[1]['name'] == 'comp25'
-        assert loaded_ids[1]['resolution'] == 1000
-
-    def test_load_same_subcomposite(self):
-        """Test loading a composite and one of it's subcomposites at the same time."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp24', 'comp25'], resolution=500)
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        assert loaded_ids[0]['name'] == 'comp24'
-        assert loaded_ids[0]['resolution'] == 500
-        assert loaded_ids[1]['name'] == 'comp25'
-        assert loaded_ids[1]['resolution'] == 500
-
-    def test_load_comp5(self):
-        """Test loading a composite that has an optional prerequisite."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp5'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp5')
-
-    def test_load_comp6(self):
-        """Test loading a composite that has an optional composite prerequisite."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp6'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp6')
-
-    def test_load_comp8(self):
-        """Test loading a composite that has a non-existent prereq."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        pytest.raises(KeyError, scene.load, ['comp8'])
-
-    def test_load_comp9(self):
-        """Test loading a composite that has a non-existent optional prereq."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp9'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp9')
-
-    def test_load_comp10(self):
-        """Test loading a composite that depends on a modified dataset."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp10'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp10')
-
-    def test_load_comp11(self):
-        """Test loading a composite that depends all wavelengths."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp11'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp11')
-
-    def test_load_comp12(self):
-        """Test loading a composite that depends all wavelengths that get modified."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp12'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp12')
-
-    def test_load_comp13(self):
-        """Test loading a composite that depends on a modified dataset where the resolution changes."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp13'])
-        loaded_ids = list(scene.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp13')
-
-    def test_load_comp14(self):
-        """Test loading a composite that updates the DataID during generation."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp14'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'comp14'
-
-    def test_load_comp15(self):
-        """Test loading a composite whose prerequisites can't be loaded.
-
-        Note that the prereq exists in the reader, but fails in loading.
-
-        """
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp15'])
-        loaded_ids = list(scene._datasets.keys())
-        assert not loaded_ids
-
-    def test_load_comp16(self):
-        """Test loading a composite whose opt prereq can't be loaded.
-
-        Note that the prereq exists in the reader, but fails in loading
-
-        """
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp16'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['name'] == 'comp16'
-
-    def test_load_comp17(self):
-        """Test loading a composite that depends on a composite that won't load."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp17'])
-        loaded_ids = list(scene._datasets.keys())
-        assert not loaded_ids
-
-    def test_load_comp18(self):
-        """Test loading a composite that depends on a incompatible area modified dataset."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt', 'fake1_highres_1.txt'], reader='fake1')
-        scene.load(['comp18'])
-        loaded_ids = list(scene._datasets.keys())
-        # depends on:
-        #   ds3
-        #   ds4 (mod1, mod3)
-        #   ds5 (mod1, incomp_areas)
-        # We should end up with ds3, ds4 (mod1, mod3), ds5 (mod1), and ds1
-        # for the incomp_areas modifier
-        assert len(loaded_ids) == 4  # the 1 dependencies
-        assert 'ds3' in scene._datasets
-        assert make_dataid(name='ds4', calibration='reflectance',
-                           modifiers=('mod1', 'mod3')) in scene._datasets
-        assert make_dataid(name='ds5', resolution=250,
-                           modifiers=('mod1',)) in scene._datasets
-
-    def test_load_comp18_2(self):
-        """Test loading a composite that depends on a incompatible area modified dataset.
-
-        Specifically a modified dataset where the modifier has optional
-        dependencies.
-
-        """
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt', 'fake1_highres_1.txt'], reader='fake1')
-        scene.load(['comp18_2'])
-        loaded_ids = list(scene._datasets.keys())
-        # depends on:
-        #   ds3
-        #   ds4 (mod1, mod3)
-        #   ds5 (mod1, incomp_areas_opt)
-        # We should end up with ds3, ds4 (mod1, mod3), ds5 (mod1), and ds1
-        # and ds2 for the incomp_areas_opt modifier
-        assert len(loaded_ids) == 5  # the 1 dependencies
-        assert 'ds3' in scene._datasets
-        assert 'ds2' in scene._datasets
-        assert make_dataid(name='ds4', calibration='reflectance',
-                           modifiers=('mod1', 'mod3')) in scene._datasets
-        assert make_dataid(name='ds5', resolution=250,
-                           modifiers=('mod1',)) in scene._datasets
-
-    def test_load_comp19(self):
-        """Test loading a composite that shares a dep with a dependency.
-
-        More importantly test that loading a dependency that depends on
-        the same dependency as this composite (a sibling dependency) and
-        that sibling dependency includes a modifier. This test makes sure
-        that the Node in the dependency tree is the exact same node.
-
-        """
-        # Check dependency tree nodes
-        # initialize the dep tree without loading the data
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene._update_dependency_tree({'comp19'}, None)
-
-        this_node = scene._dependency_tree['comp19']
-        shared_dep_id = make_dataid(name='ds5', modifiers=('res_change',))
-        shared_dep_expected_node = scene._dependency_tree[shared_dep_id]
-        # get the node for the first dep in the prereqs list of the
-        # comp13 node
-        shared_dep_node = scene._dependency_tree['comp13'].data[1][0]
-        shared_dep_node2 = this_node.data[1][0]
-        assert shared_dep_expected_node is shared_dep_node
-        assert shared_dep_expected_node is shared_dep_node2
-
-        scene.load(['comp19'])
-
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp19')
-
-    def test_load_multiple_comps(self):
-        """Test loading multiple composites."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp1', 'comp2', 'comp3', 'comp4', 'comp5', 'comp6',
-                    'comp7', 'comp9', 'comp10'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 9
-
-    def test_load_multiple_comps_separate(self):
-        """Test loading multiple composites, one at a time."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp10'])
-        scene.load(['comp9'])
-        scene.load(['comp7'])
-        scene.load(['comp6'])
-        scene.load(['comp5'])
-        scene.load(['comp4'])
-        scene.load(['comp3'])
-        scene.load(['comp2'])
-        scene.load(['comp1'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 9
-
-    def test_load_modified(self):
-        """Test loading a modified dataset."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load([make_dsq(name='ds1', modifiers=('mod1', 'mod2'))])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['modifiers'] == ('mod1', 'mod2')
-
-    def test_load_modified_with_load_kwarg(self):
-        """Test loading a modified dataset using the ``Scene.load`` keyword argument."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'], modifiers=('mod1', 'mod2'))
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0]['modifiers'] == ('mod1', 'mod2')
-
-    def test_load_multiple_modified(self):
-        """Test loading multiple modified datasets."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load([
-            make_dataid(name='ds1', modifiers=('mod1', 'mod2')),
-            make_dataid(name='ds2', modifiers=('mod2', 'mod1')),
-        ])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        for i in loaded_ids:
-            if i['name'] == 'ds1':
-                assert i['modifiers'] == ('mod1', 'mod2')
-            else:
-                assert i['name'] == 'ds2'
-                assert i['modifiers'] == ('mod2', 'mod1')
-
-    def test_load_dataset_after_composite(self):
-        """Test load composite followed by other datasets."""
-        from satpy.readers.yaml_reader import FileYAMLReader
-        from satpy.tests.utils import FakeCompositor
-        load_mock = spy_decorator(FileYAMLReader.load)
-        comp_mock = spy_decorator(FakeCompositor.__call__)
-        with mock.patch.object(FileYAMLReader, 'load', load_mock),\
-                mock.patch.object(FakeCompositor, '__call__', comp_mock):
-            lmock = load_mock.mock
-            scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-            scene.load(['comp3'])
-            assert lmock.call_count == 1
-            scene.load(['ds1'])
-            assert lmock.call_count == 2
-            scene.load(['ds1'])
-            # we should only load from the file twice
-            assert lmock.call_count == 2
-            # we should only generate the composite once
-            assert comp_mock.mock.call_count == 1
-            loaded_ids = list(scene._datasets.keys())
-            assert len(loaded_ids) == 2
-
-    def test_load_dataset_after_composite2(self):
-        """Test load complex composite followed by other datasets."""
-        from satpy.readers.yaml_reader import FileYAMLReader
-        from satpy.tests.utils import FakeCompositor, FakeModifier
-        load_mock = spy_decorator(FileYAMLReader.load)
-        comp_mock = spy_decorator(FakeCompositor.__call__)
-        mod_mock = spy_decorator(FakeModifier.__call__)
-        with mock.patch.object(FileYAMLReader, 'load', load_mock), \
-             mock.patch.object(FakeCompositor, '__call__', comp_mock), \
-             mock.patch.object(FakeModifier, '__call__', mod_mock):
-            lmock = load_mock.mock
-            scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-            scene.load(['comp10'])
-            assert lmock.call_count == 1
-            loaded_ids = list(scene._datasets.keys())
-            assert len(loaded_ids) == 1
-            with mock.patch.object(scene, '_generate_composites_nodes_from_loaded_datasets',
-                                   wraps=scene._generate_composites_nodes_from_loaded_datasets) as m:
-                scene.load(['ds1'])
-                assert lmock.call_count == 2
-                loaded_ids = list(scene._datasets.keys())
-                assert len(loaded_ids) == 2
-                # this is the unmodified ds1
-                assert make_dataid(
-                    name='ds1', resolution=250, calibration='reflectance', modifiers=tuple()
-                ) in loaded_ids
-                # m.assert_called_once_with(set([scene._dependency_tree['ds1']]))
-                m.assert_called_once_with(set())
-            with mock.patch.object(scene, '_generate_composites_nodes_from_loaded_datasets',
-                                   wraps=scene._generate_composites_nodes_from_loaded_datasets) as m:
-                scene.load(['ds1'])
-                assert lmock.call_count == 2
-                loaded_ids = list(scene._datasets.keys())
-                assert len(loaded_ids) == 2
-                # this is the unmodified ds1
-                assert make_dataid(
-                    name='ds1', resolution=250, calibration='reflectance', modifiers=tuple()
-                ) in loaded_ids
-                m.assert_called_once_with(set())
-            # we should only generate the comp10 composite once but comp2 was also generated
-            assert comp_mock.mock.call_count == 1 + 1
-            # Create the modded ds1 at comp10, then load the umodified version
-            # again
-            assert mod_mock.mock.call_count == 1
-            loaded_ids = list(scene._datasets.keys())
-            assert len(loaded_ids) == 2
-
-    def test_load_comp20(self):
-        """Test loading composite with optional modifier dependencies."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp20'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp20')
-
-    def test_load_comp21(self):
-        """Test loading composite with bad optional modifier dependencies."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp21'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp21')
-
-    def test_load_comp22(self):
-        """Test loading composite with only optional modifier dependencies."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp22'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='comp22')
-
-    def test_load_green(self):
-        """Test loading ahi_green."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ahi_green'])
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 1
-        assert loaded_ids[0] == make_cid(name='ahi_green')
-
-    def test_no_generate_comp10(self):
-        """Test generating a composite after loading."""
-        # it is fine that an optional prereq doesn't exist
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['comp10'], generate=False)
-        assert any(ds_id['name'] == 'comp10' for ds_id in scene._wishlist)
-        assert 'comp10' not in scene._datasets
-        # two dependencies should have been loaded
-        assert len(scene._datasets) == 2
-        assert len(scene.missing_datasets) == 1
-
-        scene._generate_composites_from_loaded_datasets()
-        assert any(ds_id['name'] == 'comp10' for ds_id in scene._wishlist)
-        assert 'comp10' in scene._datasets
-        assert not scene.missing_datasets
-
-    def test_modified_with_wl_dep(self):
-        """Test modifying a dataset with a modifier with modified deps.
-
-        More importantly test that loading the modifiers dependency at the
-        same time as the original modified dataset that the dependency tree
-        nodes are unique and that DataIDs.
-
-        """
-        from satpy.dataset.dataid import WavelengthRange
-
-        # Check dependency tree nodes
-        # initialize the dep tree without loading the data
-        ds1_mod_id = make_dsq(name='ds1', modifiers=('mod_wl',))
-        ds3_mod_id = make_dsq(name='ds3', modifiers=('mod_wl',))
-
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene._update_dependency_tree({ds1_mod_id, ds3_mod_id}, None)
-
-        ds1_mod_node = scene._dependency_tree[ds1_mod_id]
-        ds3_mod_node = scene._dependency_tree[ds3_mod_id]
-        ds1_mod_dep_node = ds1_mod_node.data[1][1]
-        ds3_mod_dep_node = ds3_mod_node.data[1][1]
-        # mod_wl depends on the this node:
-        ds6_modded_node = scene._dependency_tree[make_dataid(name='ds6', modifiers=('mod1',))]
-        # this dep should be full qualified with name and wavelength
-        assert ds6_modded_node.name['name'] is not None
-        assert isinstance(ds6_modded_node.name['wavelength'], WavelengthRange)
-        # the node should be shared between everything that uses it
-        assert ds1_mod_dep_node is ds3_mod_dep_node
-        assert ds1_mod_dep_node is ds6_modded_node
-
-        # it is fine that an optional prereq doesn't exist
-        scene.load([ds1_mod_id, ds3_mod_id])
-
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        assert ds1_mod_id in scene._datasets
-        assert ds3_mod_id in scene._datasets
-
-    def test_load_comp11_and_23(self):
-        """Test loading two composites that depend on similar wavelengths."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        # mock the available comps/mods in the compositor loader
-        avail_comps = scene.available_composite_ids()
-        assert make_cid(name='comp11') in avail_comps
-        assert make_cid(name='comp23') in avail_comps
-        # it is fine that an optional prereq doesn't exist
-        scene.load(['comp11', 'comp23'])
-        comp11_node = scene._dependency_tree['comp11']
-        comp23_node = scene._dependency_tree['comp23']
-        assert comp11_node.data[1][-1].name['name'] == 'ds10'
-        assert comp23_node.data[1][0].name['name'] == 'ds8'
-        loaded_ids = list(scene._datasets.keys())
-        assert len(loaded_ids) == 2
-        assert 'comp11' in scene
-        assert 'comp23' in scene
-
-    def test_load_too_many(self):
-        """Test dependency tree if too many reader keys match."""
-        scene = Scene(filenames=['fake3_1.txt'], reader='fake3')
-        avail_comps = scene.available_composite_ids()
-        # static image => 1
-        assert len(avail_comps) == 1
-        pytest.raises(KeyError, scene.load, [0.21])
-
-    def test_available_comps_no_deps(self):
-        """Test Scene available composites when composites don't have a dependency."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        all_comp_ids = scene.available_composite_ids()
-        assert make_cid(name='static_image') in all_comp_ids
-        available_comp_ids = scene.available_composite_ids()
-        assert make_cid(name='static_image') in available_comp_ids
-
-    def test_available_when_sensor_none_in_preloaded_dataarrays(self):
-        """Test Scene available composites when existing loaded arrays have sensor set to None.
-
-        Some readers or composites (ex. static images) don't have a sensor and
-        developers choose to set it to `None`. This test makes sure this
-        doesn't break available composite IDs.
-
-        """
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene['my_data'] = _data_array_none_sensor("my_data")
-        available_comp_ids = scene.available_composite_ids()
-        assert make_cid(name='static_image') in available_comp_ids
-
-    def test_load_when_sensor_none_in_preloaded_dataarrays(self):
-        """Test Scene loading when existing loaded arrays have sensor set to None.
-
-        Some readers or composites (ex. static images) don't have a sensor and
-        developers choose to set it to `None`. This test makes sure this
-        doesn't break loading.
-
-        """
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene['my_data'] = _data_array_none_sensor("my_data")
-        scene.load(["static_image"])
-        assert "static_image" in scene
-        assert "my_data" in scene
-
-    def test_compute_pass_through(self):
-        """Test pass through of xarray compute."""
-        import numpy as np
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'])
-        scene = scene.compute()
-        assert isinstance(scene['ds1'].data, np.ndarray)
-
-    def test_persist_pass_through(self):
-        """Test pass through of xarray persist."""
-        from dask.array.utils import assert_eq
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'])
-        scenep = scene.persist()
-        assert_eq(scene['ds1'].data, scenep['ds1'].data)
-        assert set(scenep['ds1'].data.dask).issubset(scene['ds1'].data.dask)
-        assert len(scenep["ds1"].data.dask) == scenep['ds1'].data.npartitions
-
-    def test_chunk_pass_through(self):
-        """Test pass through of xarray chunk."""
-        scene = Scene(filenames=['fake1_1.txt'], reader='fake1')
-        scene.load(['ds1'])
-        scene = scene.chunk(chunks=2)
-        assert scene['ds1'].data.chunksize == (2, 2)
-
-
-def _data_array_none_sensor(name: str) -> xr.DataArray:
-    """Create a DataArray with sensor set to ``None``."""
-    return xr.DataArray(
-        da.zeros((2, 2)),
-        attrs={
-            "name": name,
-            "sensor": None,
-        })
-
-
+@pytest.mark.usefixtures("include_test_etc")
 class TestSceneResampling:
     """Test resampling a Scene to another Scene object."""
-
-    def setup_method(self):
-        """Set config_path to point to test 'etc' directory."""
-        self.old_config_path = satpy.config.get('config_path')
-        satpy.config.set(config_path=[TEST_ETC_DIR])
-
-    def teardown_method(self):
-        """Restore previous 'config_path' setting."""
-        satpy.config.set(config_path=self.old_config_path)
 
     def _fake_resample_dataset(self, dataset, dest_area, **kwargs):
         """Return copy of dataset pretending it was resampled."""
